moved hotspot.ts
