+++ packages/model-viewer/src/test/index.ts
@@ -30,6 +30,7 @@ import './three-components/SmoothControls-spec.js';
 import './three-components/TextureUtils-spec.js';
 import './three-components/CachingGLTFLoader-spec.js';
 import './three-components/ModelUtils-spec.js';
+import './three-components/Hotspot-spec.js';
 import './utilities/animation-spec.js';
 import './utilities/cache-eviction-policy-spec.js';
 import './utilities/focus-visible-spec.js';
@@ -37,7 +38,6 @@ import './utilities/progress-tracker-spec.js';
 import './utilities/timer-spec.js';
 import './features/animation-spec.js';
 import './features/annotation-spec.js';
-import './features/annotation/hotspot-spec.js';
 import './features/staging-spec.js';
 import './features/controls-spec.js';
 import './features/environment-spec.js';
