+++ packages/model-viewer/src/test/three-components/Hotspot-spec.ts
@@ -12,8 +12,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import {Hotspot, HotspotVisibilityDetails} from '../../../features/annotation/hotspot.js';
-import {waitForEvent} from '../../helpers.js';
+import {Hotspot, HotspotVisibilityDetails} from '../../three-components/Hotspot.js';
+import {waitForEvent} from '../helpers.js';
 
 const expect = chai.expect;
