+++ packages/model-viewer/src/three-components/Hotspot.ts
@@ -15,8 +15,8 @@
 
 import {Vector3} from 'three';
 import {CSS2DObject} from 'three/examples/jsm/renderers/CSS2DRenderer.js';
-import {normalizeUnit} from '../../styles/conversions.js';
-import {NumberNode, parseExpressions} from '../../styles/parsers.js';
+import {normalizeUnit} from '../styles/conversions.js';
+import {NumberNode, parseExpressions} from '../styles/parsers.js';
 
 export interface HotspotVisibilityDetails {
   visible: boolean;
